Define relationships between Club and Bag/User. Create a ClubFactory class and update the club_table migration.

+++ database/migrations/2024_04_09_130408_create_clubs_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->enum('type', Club::TYPE_OPTIONS)->nullable();
             $table->string('make')->nullable();
             $table->string('model')->nullable();
-            $table->tinyInteger('average_carry')->nullable();
-            $table->tinyInteger('average_total')->nullable();
+            $table->bigInteger('average_carry')->nullable();
+            $table->bigInteger('average_total')->nullable();
             $table->timestamps();
         });
     }
